train_4h_final.py
# train_4h_final.py — FINAL TRAINING CODE (matches live bot 100%)
# Run this ONCE to generate all model files

import numpy as np
import pandas as pd
import torch
import torch.nn as nn
import torch.optim as optim
from sklearn.preprocessing import StandardScaler
from datasets import load_dataset
import os
from datetime import datetime, timedelta

print("=== TRAINING FINAL 4H BTC REGIME BOT ===")

# Load 1m data → resample to 4h
def get_4h_data():
    print("Loading WinkingFace/CryptoLM-Bitcoin-BTC-USDT...")
    ds = load_dataset("WinkingFace/CryptoLM-Bitcoin-BTC-USDT", split="train")
    df = ds.to_pandas()
    df['timestamp'] = pd.to_datetime(df['timestamp'])
    df = df.set_index('timestamp')[['open','high','low','close','volume']].astype(float)
    df = df.sort_index()
    df_4h = df.resample('4h').agg({
        'open': 'first', 'high': 'max', 'low': 'min',
        'close': 'last', 'volume': 'sum'
    }).dropna()
    print(f"Resampled: {len(df_4h)} 4h candles from {df_4h.index[0]} to {df_4h.index[-1]}")
    return df_4h

btc = get_4h_data()
btc['close'] = btc['close'].ffill()
btc['return'] = np.log(btc['close'] / btc['close'].shift(1))

# Synthetic funding (live uses real)
np.random.seed(42)
btc['funding_rate'] = np.tanh(np.cumsum(np.random.normal(0, 0.0003, len(btc)))) * 0.0005
btc = btc.dropna()

# EKF
def run_ekf(p):
    n = len(p)
    x = np.zeros((n, 3))
    x[0] = [p[0], 0, -5]
    P = np.eye(3)
    Q = np.diag([1e-3, 1e-5, 1e-7])
    R = 0.3**2
    level = np.zeros(n)
    velocity = np.zeros(n)
    for t in range(1, n):
        F = np.array([[1, 0.1667, 0], [0, 1, 0], [0, 0, 1]])
        x_pred = F @ x[t-1]
        P_pred = F @ P @ F.T + Q
        y = p[t] - x_pred[0]
        S = P_pred[0,0] + R
        K = P_pred[:,0] / S
        x[t] = x_pred + K * y
        P = (np.eye(3) - np.outer(K, [1,0,0])) @ P_pred
        level[t] = x[t][0]
        velocity[t] = x[t][1]
    return level, velocity

print("Running EKF...")
btc['level'], btc['velocity'] = run_ekf(btc['close'].values)

# Features
window = 180
features, targets = [], []
for i in range(window, len(btc)-1):
    features.append([
        btc['level'].iloc[i],
        btc['velocity'].iloc[i],
        btc['funding_rate'].iloc[i],
        btc['return'].iloc[i-24:i+1].mean(),
        btc['close'].iloc[i] / btc['close'].iloc[i-window:i].mean() - 1
    ])
    targets.append(btc['return'].iloc[i+1])

X = np.array(features)
y = np.array(targets)
split = int(0.8 * len(X))

scaler = StandardScaler()
X_train_s = scaler.fit_transform(X[:split])
X_test_s = scaler.transform(X[split:])

# FFNN
class FFNN(nn.Module):
    def __init__(self):
        super().__init__()
        self.net = nn.Sequential(
            nn.Linear(5, 128), nn.ReLU(), nn.Dropout(0.3),
            nn.Linear(128, 64), nn.ReLU(), nn.Dropout(0.3),
            nn.Linear(64, 1)
        )
    def forward(self, x): return self.net(x)

model = FFNN()
opt = optim.AdamW(model.parameters(), lr=0.001)
crit = nn.MSELoss()
print("Training FFNN...")
for epoch in range(500):
    opt.zero_grad()
    pred = model(torch.FloatTensor(X_train_s))
    loss = crit(pred, torch.FloatTensor(y[:split]).unsqueeze(1))
    loss.backward()
    opt.step()
    if epoch % 100 == 0:
        print(f"Epoch {epoch}: {loss.item():.6f}")

# HyperDUM — 4 Regimes (robust version)
D = 2048
np.random.seed(42)
projector = np.random.randn(5, D)
projector /= np.linalg.norm(projector, axis=0, keepdims=True)

feature_dates = btc.index[window:window + split]
velocity_series = btc['velocity'].iloc[window:window + split]
funding_series = btc['funding_rate'].iloc[window:window + split]
volume_series = btc['volume'].iloc[window:window + split]

halving_date = pd.Timestamp("2024-04-19")

# Robust masks
halving_window = (feature_dates >= halving_date - timedelta(days=180)) & (feature_dates < halving_date)
low_vol = np.abs(velocity_series) < velocity_series.quantile(0.4)
positive_funding = funding_series > 0
halving_mask = halving_window & low_vol & positive_funding

normal_mask = (feature_dates > halving_date + timedelta(days=180)) & \
              (np.abs(velocity_series) < velocity_series.quantile(0.6))

high_vol = np.abs(velocity_series) > velocity_series.quantile(0.95)
high_volume = volume_series > volume_series.quantile(0.9)
extreme_funding = np.abs(funding_series) > funding_series.abs().quantile(0.9)
blowoff_mask = high_vol & high_volume & extreme_funding

bear_mask = ~(normal_mask | halving_mask | blowoff_mask)

# Convert to numpy
for mask in ['normal_mask', 'halving_mask', 'blowoff_mask', 'bear_mask']:
    if hasattr(locals()[mask], 'values'):
        locals()[mask] = locals()[mask].values

def make_memory(X_subset):
    if len(X_subset) == 0:
        return np.zeros(D)
    hd = np.sign(X_subset @ projector)
    return np.sign(np.sum(hd, axis=0))

memory_normal   = make_memory(X_train_s[normal_mask])
memory_halving  = make_memory(X_train_s[halving_mask])
memory_blowoff  = make_memory(X_train_s[blowoff_mask])
memory_bear     = make_memory(X_train_s[bear_mask])

# Save
os.makedirs("model", exist_ok=True)
torch.save(model, "model/btc_4h_model.pth")
torch.save(scaler, "model/scaler_4h.pth")
np.save("model/projector_4h.npy", projector)
np.save("model/memory_normal.npy", memory_normal)
np.save("model/memory_halving.npy", memory_halving)
np.save("model/memory_blowoff.npy", memory_blowoff)
np.save("model/memory_bear.npy", memory_bear)

print("\nTRAINING COMPLETE — ALL FILES SAVED")
print("Run live_bot_sms.py → you're live with SMS alerts")
