requirements.txt
torch
numpy
pandas
scikit-learn
python-dotenv
requests
yfinance